Delete files after edit, bug fixes
Delete files from server after editing and some minor fixes

+++ image/edit_2.py
@@ -6,63 +6,89 @@ from PIL import Image,ImageEnhance,ImageDraw,ImageFilter,ImageOps
 import numpy as np
 import os
 import cv2
+import shutil
+
 
 async def circle(client, message):
-    media = message
-    download_location = "./DOWNLOADS" + "/" + str(message.from_user.id) + ".jpg"
-    msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
-    a  =   await client.download_media(
-           message=media.reply_to_message,
-           file_name=download_location
-        )
-    await msg.edit("Processing Image...")
-    img=Image.open(a).convert("RGB")
-    npImage = np.array(img)
-    h,w = img.size
-    alpha = Image.new('L', img.size,0)
-    draw = ImageDraw.Draw(alpha)
-    draw.pieslice([0,0,h,w],0,360,fill=255)
-    npAlpha=np.array(alpha)
-    npImage=np.dstack((npImage,npAlpha))
-    Image.fromarray(npImage).save('circle.png')
-    await msg.delete()
-    await message.reply_chat_action("upload_photo")
-    await message.reply_to_message.reply_photo("circle.png", quote=True)
-
-
+    userid = str(message.from_user.id)
+    if not os.path.isdir(f"./DOWNLOADS/{userid}"):
+        os.mkdir(f"./DOWNLOADS/{userid}")
+    download_location = "./DOWNLOADS" + "/" + userid + ".jpg"
+    edit_img_loc = "./DOWNLOADS" + "/" + userid + "/" + "circle.png"
+    if not message.reply_to_message.empty:
+        msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
+        a  =   await client.download_media(
+               message=message.reply_to_message,
+               file_name=download_location
+            )
+        await msg.edit("Processing Image...")
+        img=Image.open(a).convert("RGB")
+        npImage = np.array(img)
+        h,w = img.size
+        alpha = Image.new('L', img.size,0)
+        draw = ImageDraw.Draw(alpha)
+        draw.pieslice([0,0,h,w],0,360,fill=255)
+        npAlpha=np.array(alpha)
+        npImage=np.dstack((npImage,npAlpha))
+        Image.fromarray(npImage).save(edit_img_loc)
+        await message.reply_chat_action("upload_photo")
+        await message.reply_to_message.reply_photo(edit_img_loc, quote=True)
+        await msg.delete()
+    else:
+        await message.reply_text("Why did you delete that??")
+    try:
+        shutil.rmtree(f"./DOWNLOADS/{userid}")
+    except:
+        pass
 
 async def sticker(client, message):
-    media = message
-    download_location = "./DOWNLOADS" + "/" + str(message.from_user.id) + ".jpg"
-    msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
-    a  =   await client.download_media(
-           message=media.reply_to_message,
-           file_name=download_location
-        )
-    await msg.edit("Processing Image...")
-    os.rename(a,"sticker.webp")
-    await msg.delete()
-    await  message.reply_to_message.reply_sticker("sticker.webp", quote=True)
-
+    userid = str(message.from_user.id)
+    if not os.path.isdir(f"./DOWNLOADS/{userid}"):
+        os.mkdir(f"./DOWNLOADS/{userid}")
+    download_location = "./DOWNLOADS" + "/" + userid + ".jpg"
+    edit_img_loc = "./DOWNLOADS" + "/" + userid + "/" + "sticker.webp"
+    if not message.reply_to_message.empty:
+        msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
+        a  =   await client.download_media(
+               message=message.reply_to_message,
+               file_name=download_location
+            )
+        await msg.edit("Processing Image...")
+        os.rename(a,edit_img_loc)
+        await  message.reply_to_message.reply_sticker(edit_img_loc, quote=True)
+        await msg.delete()
+    else:
+        await message.reply_text("Why did you delete that??")
+    try:
+        shutil.rmtree(f"./DOWNLOADS/{userid}")
+    except:
+        pass
 
 async def contrast(client, message):
-    media = message
-    download_location = "./DOWNLOADS" + "/" + str(message.from_user.id) + ".jpg"
-    msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
-    a  =   await client.download_media(
-           message=media.reply_to_message,
-           file_name=download_location
-        )
-    await msg.edit("Processing Image...")
-    image=Image.open(a)
-    contrast=ImageEnhance.Contrast(image)
-    contrast.enhance(1.5).save('contrast.jpg')
-    await msg.delete()
-    
-    await message.reply_chat_action("upload_photo")
-  
-    await message.reply_to_message.reply_photo("contrast.jpg", quote=True)
-
+    userid = str(message.from_user.id)
+    if not os.path.isdir(f"./DOWNLOADS/{userid}"):
+        os.mkdir(f"./DOWNLOADS/{userid}")
+    download_location = "./DOWNLOADS" + "/" + userid + ".jpg"
+    edit_img_loc = "./DOWNLOADS" + "/" + userid + "/" + "contrast.jpg"
+    if not message.reply_to_message.empty:
+        msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
+        a  =   await client.download_media(
+               message=message.reply_to_message,
+               file_name=download_location
+            )
+        await msg.edit("Processing Image...")
+        image=Image.open(a)
+        contrast=ImageEnhance.Contrast(image)
+        contrast.enhance(1.5).save(edit_img_loc)
+        await message.reply_chat_action("upload_photo")  
+        await message.reply_to_message.reply_photo(edit_img_loc, quote=True)
+        await msg.delete()
+    else:
+        await message.reply_text("Why did you delete that??")
+    try:
+        shutil.rmtree(f"./DOWNLOADS/{userid}")
+    except:
+        pass
 
 
 def sepia(img):
@@ -87,45 +113,63 @@ def sepia(img):
     return new_img
 
 async def sepia_mode(client, message):
-    media = message
-    download_location = "./DOWNLOADS" + "/" + str(message.from_user.id) + ".jpg"
-    msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
-    a  =   await client.download_media(
-           message=media.reply_to_message,
-           file_name=download_location
-        )
-    await msg.edit("Processing Image...")
-    image=Image.open(a)
-    new_img = sepia(image)
-    new_img.save("sepia.jpg")
-    await msg.delete()
-    await message.reply_chat_action("upload_photo")
-    await message.reply_to_message.reply_photo("sepia.jpg", quote=True)
-
+    userid = str(message.from_user.id)
+    if not os.path.isdir(f"./DOWNLOADS/{userid}"):
+        os.mkdir(f"./DOWNLOADS/{userid}")
+    download_location = "./DOWNLOADS" + "/" + userid + ".jpg"
+    edit_img_loc = "./DOWNLOADS" + "/" + userid + "/" + "sepia.jpg"
+    if not message.reply_to_message.empty:
+        msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
+        a  =   await client.download_media(
+               message=message.reply_to_message,
+               file_name=download_location
+            )
+        await msg.edit("Processing Image...")
+        image=Image.open(a)
+        new_img = sepia(image)
+        new_img.save(edit_img_loc)
+        await message.reply_chat_action("upload_photo")
+        await message.reply_to_message.reply_photo(edit_img_loc, quote=True)
+        await msg.delete()
+    else:
+        await message.reply_text("Why did you delete that??")
+    try:
+        shutil.rmtree(f"./DOWNLOADS/{userid}")
+    except:
+        pass
 
 def dodgeV2(x, y):
        return cv2.divide(x, 255 - y, scale=256)
 
 
 async def pencil(client, message):
-    media = message
-    download_location = "./DOWNLOADS" + "/" + str(message.from_user.id) + ".jpg"
-    msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
-    a  =   await client.download_media(
-           message=media.reply_to_message,
-           file_name=download_location
-        )
-    await msg.edit("Processing Image...")
-    img = cv2.imread(a)
-    img_gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-    img_invert = cv2.bitwise_not(img_gray)
-    img_smoothing = cv2.GaussianBlur(img_invert, (21, 21),sigmaX=0, sigmaY=0)
-    final_img = dodgeV2(img_gray, img_smoothing)
-    cv2.imwrite("pencil.jpg", final_img)
-    await msg.delete()
-    await message.reply_chat_action("upload_photo")
-    await message.reply_to_message.reply_photo("pencil.jpg", quote=True)
-
+    userid = str(message.from_user.id)
+    if not os.path.isdir(f"./DOWNLOADS/{userid}"):
+        os.mkdir(f"./DOWNLOADS/{userid}")
+    download_location = "./DOWNLOADS" + "/" + userid + ".jpg"
+    edit_img_loc = "./DOWNLOADS" + "/" + userid + "/" + "pencil.jpg"
+    if not message.reply_to_message.empty:
+        msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
+        a  =   await client.download_media(
+               message=message.reply_to_message,
+               file_name=download_location
+            )
+        await msg.edit("Processing Image...")
+        img = cv2.imread(a)
+        img_gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+        img_invert = cv2.bitwise_not(img_gray)
+        img_smoothing = cv2.GaussianBlur(img_invert, (21, 21),sigmaX=0, sigmaY=0)
+        final_img = dodgeV2(img_gray, img_smoothing)
+        cv2.imwrite(edit_img_loc, final_img)
+        await message.reply_chat_action("upload_photo")
+        await message.reply_to_message.reply_photo(edit_img_loc, quote=True)
+        await msg.delete()
+    else:
+        await message.reply_text("Why did you delete that??")
+    try:
+        shutil.rmtree(f"./DOWNLOADS/{userid}")
+    except:
+        pass
 
 
 
@@ -141,23 +185,34 @@ def color_quantization(img, k):
 
 
 async def cartoon(client, message):
-    media = message
-    download_location = "./DOWNLOADS" + "/" + str(message.from_user.id) + ".jpg"
-    msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
-    a  =   await client.download_media(
-           message=media.reply_to_message,
-           file_name=download_location
-        )
-    await msg.edit("Processing Image...")
-    img = cv2.imread(a)
-    edges = cv2.Canny(img, 100, 200)
-    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
-    edges = cv2.adaptiveThreshold(gray, 255, cv2.ADAPTIVE_THRESH_MEAN_C, cv2.THRESH_BINARY, 9, 5)
-    color = cv2.bilateralFilter(img, d=9, sigmaColor=200,sigmaSpace=200)
-   
-    cartoon = cv2.bitwise_and(color, color, mask=edges) 
-    img_1 = color_quantization(img, 7)
-    cv2.imwrite("kang.jpg",img_1)
-    await msg.delete()
-    await message.reply_chat_action("upload_photo")
-    await message.reply_to_message.reply_photo("kang.jpg", quote=True)
+    userid = str(message.from_user.id)
+    if not os.path.isdir(f"./DOWNLOADS/{userid}"):
+        os.mkdir(f"./DOWNLOADS/{userid}")
+    download_location = "./DOWNLOADS" + "/" + userid + ".jpg"
+    edit_img_loc = "./DOWNLOADS" + "/" + userid + "/" + "kang.jpg"
+    if not message.reply_to_message.empty:
+        msg = await message.reply_to_message.reply_text("Downloading image", quote=True)
+        a  =   await client.download_media(
+               message=message.reply_to_message,
+               file_name=download_location
+            )
+        await msg.edit("Processing Image...")
+        img = cv2.imread(a)
+        edges = cv2.Canny(img, 100, 200)
+        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+        edges = cv2.adaptiveThreshold(gray, 255, cv2.ADAPTIVE_THRESH_MEAN_C, cv2.THRESH_BINARY, 9, 5)
+        color = cv2.bilateralFilter(img, d=9, sigmaColor=200,sigmaSpace=200)
+       
+        cartoon = cv2.bitwise_and(color, color, mask=edges) 
+        img_1 = color_quantization(img, 7)
+        cv2.imwrite(edit_img_loc,img_1)
+        await message.reply_chat_action("upload_photo")
+        await message.reply_to_message.reply_photo(edit_img_loc, quote=True)
+        await msg.delete()
+    else:
+        await message.reply_text("Why did you delete that??")
+    try:
+        shutil.rmtree(f"./DOWNLOADS/{userid}")
+    except:
+        pass
+
